impl common field component
