+++ src/components/TargetCountCard/TargetCountsCard.tsx
@@ -1,7 +1,6 @@
+"use client";
 import React from 'react';
-import { TextField } from '@mui/material';
-import { T } from '@util/T';
-import { rowLabel, rowSep } from '@components/StatsCard/StatsCard';
+import { FieldCells, type FieldDef } from '@components/FieldCells/FieldCells';
 
 interface TargetCountsCardProps {
     options: {
@@ -9,39 +8,22 @@ interface TargetCountsCardProps {
         allyCount: number;
     };
     onOptionsChange: (newOptions: any) => void;
+    label?: string;
+    fields?: FieldDef[];
 }
 
-const targets = [
-    { key: 'enemyCount' as const, label: 'Enemies' },
-    { key: 'allyCount' as const, label: 'Allies' },
+const defaultFields: FieldDef[] = [
+    { key: 'enemyCount', label: 'enemies', min: 1 },
+    { key: 'allyCount', label: 'allies', min: 1 },
 ];
 
-const inputSx = { width: "100%", maxWidth: 110, '& .MuiInputBase-input': { textAlign: 'right', fontSize: '0.9rem', fontFamily: 'monospace' } };
-
-const TargetCountsCard: React.FC<TargetCountsCardProps> = ({ options, onOptionsChange }) => {
-    const handleChange = (field: string, value: number) => {
-        onOptionsChange((prev: any) => ({ ...prev, [field]: value }));
-    };
-
-    return (
-        <div style={{ display: "grid", gridTemplateColumns: "max-content 1px auto", gap: "4px 10px", alignItems: "center" }}>
-            {targets.map(({ key, label }) => (
-                <React.Fragment key={key}>
-                    <span style={rowLabel}>{T(label)}</span>
-                    {rowSep}
-                    <TextField
-                        type="text"
-                        size="small"
-                        variant="standard"
-                        value={options[key]}
-                        onChange={(e) => handleChange(key, Math.max(1, Number(e.target.value)))}
-                        slotProps={{ htmlInput: { min: 1, inputMode: 'numeric' } }}
-                        sx={inputSx}
-                    />
-                </React.Fragment>
-            ))}
-        </div>
-    );
-};
+const TargetCountsCard: React.FC<TargetCountsCardProps> = ({ options, onOptionsChange, label, fields = defaultFields }) => (
+    <FieldCells
+        fields={fields}
+        options={options}
+        onOptionsChange={onOptionsChange}
+        label={label}
+    />
+);
 
 export default TargetCountsCard;
